[TEST] Extract shared functions
